% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app.R
\name{app}
\alias{app}
\title{Runs the GPCAH Surveillance of Agriculture Injuries in Iowa Web Application}
\usage{
app(ui = agstudy1app::ui, server = agstudy1app::server)
}
\arguments{
\item{ui}{The UI definition of the app (for example, a call to fluidPage()
with nested controls). Default is \code{\link[=ui]{ui()}}.}

\item{server}{A function with three parameters: input, output, and session.
The function is called once for each session ensuring that each app is
independent. Default is \code{\link[=server]{server()}}.}
}
\description{
Runs a local version of the web application using \code{\link[shiny:shinyApp]{shiny::shinyApp()}}.
}
\seealso{
\code{\link[shiny:shinyApp]{shiny::shinyApp()}}, \code{\link[shiny:runApp]{shiny::runApp()}}
}
