% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covv}
\alias{covv}
\title{Variance-covariance matrix from logistic regression model}
\format{
\subsection{\code{covv}}{

A data frame with 15 row and 15 columns (variables):
\describe{
\item{int}{Intercept}
\item{s_female}{Sex is female}
\item{c_cut}{Cause of injury is related to cuts...}
\item{c_falls}{Cause of injury is related to falls...}
\item{c_fire}{Cause of injury is related to fire...}
\item{c_mvt}{Cause of injury is related to MVT...}
\item{c_mach}{Cause of injury is missing}
\item{c_nature}{Cause of injury is related to nature...}
\item{n_amp}{Nature of injury is related to amputations...}
\item{n_burns}{Nature of injury is related to burns...}
\item{n_fx}{Nature of injury is related to factures...}
\item{n_miss}{Nature of injury is missing}
\item{n_open}{Nature of injury is related to open wounds...}
\item{age}{Age}
}
}
}
\usage{
covv
}
\description{
The variance-covariance matrix from the logistic regression model used to
model the probability that a record is from the Iowa's Workers' Compensation
database.
}
\keyword{datasets}
