% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iwc_itr_only}
\alias{iwc_itr_only}
\title{Iowa's Workers' Compensation and Trauma Registry data}
\format{
\subsection{\code{iwc_itr_only}}{

A data frame with 5,817 rows and 6 columns:
\describe{
\item{obs_final}{The observation number from the combined data}
\item{sex}{Sex of record with F denoting female and M denoting male}
\item{age}{Age of recrod}
\item{cause_ov}{Numeric code for cause of injury}
\item{nature_ov}{Numeric code for nature of injury}
\item{match}{Numeric code for origin location of record with 0 denoting ITR and 1 denoting IWC}
}
}
}
\usage{
iwc_itr_only
}
\description{
A subset and modified version of the combined data from Iowa's Workers'
Compensation (IWC) database and Iowa's Trauma Registry (ITR) database.
}
\keyword{datasets}
