% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch.R
\name{launch}
\alias{launch}
\title{Runs the GPCAH Surveillance of Agriculture Injuries in Iowa Web Application}
\usage{
launch()
}
\description{
Runs a local version of the web application using \code{\link[shiny:shinyApp]{shiny::shinyApp()}}
and \code{\link[shiny:runApp]{shiny::runApp()}}.
}
\details{
This will only execute if the current working directory contains the
\emph{Web Application User Guide} (\code{"./docs/app_guide.html"}) and the \emph{Modeling
Framework} (\code{"./docs/modeling_overview.html"}).
}
\examples{
if (interactive()) {
 agstudy1app::launch()
}

}
\seealso{
\code{\link[shiny:shinyApp]{shiny::shinyApp()}}, \code{\link[shiny:runApp]{shiny::runApp()}}
}
