% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pred_prob.R
\name{pred_prob}
\alias{pred_prob}
\title{Predicted probabilities for Iowa's Workers' Compensation and Iowa's Trauma
Registry databases}
\usage{
pred_prob(sex = "NULL", cause = "NULL", nature = "NULL", age = NULL)
}
\arguments{
\item{sex}{A character value denoting the sex of the observation of interest.
Valid parameter values are \code{"Female"}, \code{"Male"}, or \code{"NULL"} with the default
being \code{"NULL"}.}

\item{cause}{A character value denoting the cause of injury of the
observation of interest. Valid parameter values are \code{"MVT"}, \code{"Falls"},
\code{"Fire/burn"}, \code{"Cut/pierces or struck by/against"}, \code{"Machinery"},
\code{"Natural/Environment"}, \code{"Missing/unspecified"}, \code{"Other"}, or \code{"NULL"} with
the default being \code{"NULL"}.}

\item{nature}{A character value denoting the nature of injury of the
observation of interest. Valid parameter values are
\code{"Fractures/dislocation/sprains/strains"},
\code{"Open Wound/superficial/contusion"}, \code{"Burns"}, \code{"Amputations"},
\code{"Missing/unspecified"}, \code{"Other"}, or \code{"NULL"} with the default value
being \code{"NULL"}.}

\item{age}{A numeric value denoting the age of the observation of interest
with the default value being \code{NULL}.}
}
\value{
A \code{data.frame} with eight variables and two rows. The variables are:
\itemize{
\item \code{source}: the location of the observation of interest,
\verb{Workers' compensation} and \verb{Trauma registry},
\item \code{sex}: the sex of the observation of interest.
\item \code{cause}: the cause of injury of the observation of interest,
\item \code{nature}: the nature of injury of the observation of interest,
\item \code{age}: the age of the observation of interest,
\item \code{probability}: the predicted probabilities that the observation of interest
is in the \verb{Workers' compensation} and \verb{Trauma registry} databases,
\item \code{lb}: the lower bound for the 95\% confidence interval of the predicted
probabilities, and
\item \code{ub}: the upper bound for the 95\% confidence interval of the predicted
probabilities.
}
}
\description{
Calculates the predicted probabilities for an observation(s) to be located in
Iowa's Workers' Compensation (IWC) and Iowa's Trauma Registry (ITR) based on
the sex, cause of injury, nature of injury, and age.
}
\details{
When a parameter is specified as the a null value
(\code{"NULL"} or \code{NULL}) the "average" parameter value is used. This is possible
because the logistic regression modeling construction used effects coding.
The "average" value is related to the constrast value for the parameter
being 0.

The parameters \code{sex}, \code{cause}, and \code{nature} use a character representation
of \code{"NULL"} because the purpose of this package and function is to help the
web application for the GPCAH Surveillance project, which is based on
\href{https://shiny.posit.co}{Shiny}.
}
\examples{
pred_prob()
pred_prob(sex = "Male", cause = "MVT", nature = "Burns", age = 45)

}
