% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Server for the GPCAH Surveillance of Agriculture Injuries in Iowa Web
Application}
\usage{
server(input, output)
}
\arguments{
\item{input}{The input from the user of the web application. These inputs are
based on the injury characteristics and demographics of the injury record of
interest.}

\item{output}{The a table to be viewed by the user.}
}
\value{
A \code{view} document created by \code{\link[shiny:renderTable]{shiny::renderTable()}}.
}
\description{
Creates the server side for the web application.
}
\details{
The server creation depends on \code{\link[shiny:renderTable]{shiny::renderTable()}} and the
\code{\link[=agstudy1app:pred_prob]{agstudy1app:pred_prob()}} functions. For more information on how the output
is generated see the \href{https://ph-ivshiny.iowa.uiowa.edu/deboonstra/agstudy1app/}{Web Application User Guide}.
}
