% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\name{ui}
\alias{ui}
\title{UI for the GPCAH Surveillance of Agriculture Injuries in Iowa Web Application}
\usage{
ui()
}
\value{
A \code{list} created by \code{\link[shiny:navbarPage]{shiny::navbarPage()}} containing HTML metadata to
be used by \link{server}.
}
\description{
Creates the UI for the web application based on navigation bar page that
includes an application user guide, an overview of the modeling framework,
and the web application.
}
\details{
The UI depends on \code{\link[shiny:navbarPage]{shiny::navbarPage()}} to create the navigation
bar page and \code{\link[shinythemes:shinytheme]{shinythemes::shinytheme()}} to define the
\href{https://bootswatch.com}{Bootswatch} theme of the application.

For this UI, there are three pages to select from. The first is titled
\emph{Web Application User Guide}, which is the user guide for the web
application. The second is an overview of modeling framework used to
calculate the predicted probabilities and is titled \emph{Modeling Framework}.
The final page is titled \emph{Web Application} and is the application itself.

The user guide and modeling framework documents are based on two HTML
fragment stored within the \verb{docs/} sub-directory. If these documents are
\strong{NOT} created then \emph{knit} the accompanied \code{.Rmd}s as one chooses. The
\code{makefile} can \emph{knit} both documents with a simple call the the make file.
}
